OpenFileGDB: implement deferred driver loading
