+++ frmts/gdalallregister.cpp
@@ -233,6 +233,9 @@ void CPL_STDCALL GDALAllRegister()
 #if defined(DEFERRED_OGDI_DRIVER)
     DeclareDeferredOGROGDIPlugin();
 #endif
+#if defined(DEFERRED_OPENFILEGDB_DRIVER)
+    DeclareDeferredOGROpenFileGDBPlugin();
+#endif
 #if defined(DEFERRED_PARQUET_DRIVER)
     DeclareDeferredOGRParquetPlugin();
 #endif
