+++ ogr/ogrsf_frmts/openfilegdb/CMakeLists.txt
@@ -15,7 +15,14 @@ add_gdal_driver(
           ogropenfilegdblayer.cpp
           ogropenfilegdblayer_write.cpp
           gdalopenfilegdbrasterband.cpp
-  PLUGIN_CAPABLE NO_DEPS STRONG_CXX_WFLAGS)
+  CORE_SOURCES
+          ogropenfilegdbdrivercore.cpp
+  PLUGIN_CAPABLE NO_DEPS
+  STRONG_CXX_WFLAGS)
+
+if(NOT TARGET ogr_OpenFileGDB)
+    return()
+endif()
 
 gdal_standard_includes(ogr_OpenFileGDB)
 target_include_directories(ogr_OpenFileGDB PRIVATE $<TARGET_PROPERTY:ogr_MEM,SOURCE_DIR>)
